adding all endpoints for backend

+++ src/main/java/com/example/demo/city/repository/CityRepository.java
@@ -3,5 +3,5 @@ package com.example.demo.city.repository;
 import com.example.demo.city.entity.CityEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CityRepository extends JpaRepository<CityEntity, Long> {
+public interface CityRepository extends JpaRepository<CityEntity, Integer> {
 }
